context: fix uint<n>_t/size_t_ptr mixup
context.py handles function input parameters.
UINT<n>_T are special because for their pointer type UINT<n>_PTR, the _T
suffix is omitted. That's not the case with all other types, including
SIZE_T, the pointer type of which is SIZE_T_PTR.

context.py accounts for the uint<n>_t special case. However, this code
path is also matches for size_t types, treating them as a special case,
too, when they are in fact not.

Signed-off-by: Johannes Holland <joh.ho@gmx.de>