Fix tpm2_pytss/config.json to work with docker image
When following the following instructions in `HACKING.md` the produced
docker image doesn't work:

```console
$ rm -rf build
$ git clean -Xf || true
$ docker build -t tpm2software/tpm2-tss-python -f .ci/Dockerfile.tpm2-tss-python .
$ docker run --rm \
  -u $(id -u):$(id -g) \
  -v "${PWD}:/workspace/tpm2-pytss" \
  --env-file .ci/docker.env \
  tpm2software/tpm2-tss-python \
  /bin/bash -c '/workspace/tpm2-pytss/.ci/docker.run'
```

The problem seems to be in the definition of the system folder passed to
tpm2_pytss to find the default FAPI configuration.
This is defined in `tpm2_pytss/config.json` which points to `/etc` when
it should be `/etc/local/usr`, or at least it's that way in the created
docker image.

Signed-off-by: Daniel Gracia Pérez <50202128+dgptha@users.noreply.github.com>