porting CFFI PR 1248 from tpm2-tss project
Ports https://github.com/tpm2-software/tpm2-tss/pull/1248
to tpm2-pytss project

building:
  - pip3 install --user -e .

Testing:
  - python3 test/esys.py

The tests pass, but a lot of work needs to be done to massage the
interface a little more. ESYS_TR's have an esys context that could be
removed out from underneath them, will cause flush on a dead esys
context pointer.

Initially the code had the Esys_Finalize on the __del__ attribute, but
since control over gc essentially determines when the Esys_Context, and
it's underlying connection to the TPM is busy, we likely want to follow
a context manager like approach with close(), __enter__ and __exit__
calls.

Release the code under BSD-3 to match with the project.

Signed-off-by: William Roberts <william.c.roberts@intel.com>