ci: run: Update test command to unittest discover
Python < 3.9 supported running tests via the `setup.py test` command.
However, in 3.9 and greater this is no longer supported. As perpetration
for supporting 3.9, we move to an alternate approach to running tests.
Thie approach is the discovery and execution of unittests via invocation
of the unittest module's discover command. More information can be found
on the unittest module's documentation page:
https://docs.python.org/3/library/unittest.html#test-discovery

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>