test: ensure tcti and executable are present for simulator
Just checking for the tpm simulator binary being present isn't enough,
becuase one can end up in the situation where the TCTI isn't available.
In that case, firing up the simulator without a way to connect will
fail. So ensure both the TCTI and executables are present, or skip that
simulator and try another one.

Signed-off-by: William Roberts <william.c.roberts@intel.com>