test: mark FAPI tests as @pytest.mark.forked
Each FAPI test needs to be run in its own subprocess because it exports the
TSS2_FAPICONF (a.k.a. FAPI_CONFIG_ENV) environment variable to point to its own
temporary FAPI configuration. Not isolating the tests in different subprocesses
leads to errors like

        if self.config is None:
            # Load the currently active fapi-config.json
            config_path = os.environ.get(FAPI_CONFIG_ENV, FAPI_CONFIG_PATH)
>           with open(config_path) as file:
E           FileNotFoundError: [Errno 2] No such file or directory: '/tmp/tmp3e2_8a4c'

because the next test tries to set up the configuration from the environment
variable, which points to the old and possibly already deleted temporary
configuration.

Mark the FAPI tests as @pytest.mark.forked from the pytest-forked/pytest-xdist
plugin (which we already require as a development dependency) to force them to
be run in individual subprocesses regardless of how the user calls pytest.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>