TPML_ make getitem smarter
Make TPML smarter by implementing the getitem in the superclass by
figuring out what field is not the count field and accessing the cdata
by that. This way we don't need to write code for each TPML subtype.

Also, add an unpacker so as we understand the data types and how folks
want to use them, we can unpack them into native types as it makes
sense.

Signed-off-by: William Roberts <william.c.roberts@intel.com>