LGTM: fix constructor calling subclass method
LGTM issued the following warning:
NEWCall to self.__setattr__ in __init__ method, which is overridden by method TPM2B_SIMPLE_OBJECT.__setattr__.

The help document for the issue is:
https://lgtm.com/rules/10030085/

LGTM's help document was clear that we should avoid this paradigm. In
this case it worked fine, but lets minimize warnings and bad patterns
when possible.

Signed-off-by: William Roberts <william.c.roberts@intel.com>