ESAPI: free types properly
Currently, CData can come from two sources:
1. allocated via ffi.new
2. allocated from a successful call to ESAPI API as returned via the
   double pointer interface.

Item one does not pose an issue, as it will use matching allocator and
deallocator routines.

Item 2 is tricky. Be default (AFAIK[1]), the pointer of the pointer will also be
handled by CFFI's free function. However, that can cause issues where
ESAPI is allocating with one implementation and cffi is deallocating
with another non-abi compatible issue. This issue can be seen with the
introduction of Esys_Free and is notably an issue on Windows.

Thus we need to explicitly, and ONLY for ESAPI allocations, set the
explicit free function. I recieved help from the mailing list[1] for cffi
on how to design this, and tested it by modifying ESAPI to print free
calls. It all seems to be working.

[1] https://groups.google.com/g/python-cffi/c/DROgzjrCwE4/m/vRn55havDAAJ

[2] Worst case I am wrong and CFFI was just leaking that pointer and now
its actually getting free'd.

Fixes: #117

Signed-off-by: William Roberts <william.c.roberts@intel.com>