docs: bail on TPM2_OBJECT init
If a default initializer for an object is a TPM2_OBJECT, the
construction of that object will occur on import of the module. However,
when building for the documentation, the CFFI layer is mocked. Mocking
the ffi.typeof and ensureing it works in each location was quite
challenging. I actually attempted to do this with by using the inspect
module, to look at the stack frames to get the expected results, inside
of the MyMagicMock implementing the CFFI calls.

It was easier and simpler to just bail on the initialization if we
detect that unittest.mock is loaded. This avoids the need for all the
complexity when building the docs, at the expense of a bail function in
the init routine.

Signed-off-by: William Roberts <william.c.roberts@intel.com>