ci: add ibmtpm server for tss2 before 3.0
Before tpm2-tss release 3.0, swtpm TCTI did not exist, so check for the
TCTI and if it exists, use swtpm, else use mssim TCTI which connects to
tpm_server.

Use version 1661 as it seems to avoid a bug in parsing the TCG Vendor
Test command. Versions before this seem to think the first paramter is a
RH ACT.

Signed-off-by: William Roberts <william.c.roberts@intel.com>