test: fix pollfd unknown size
Fixes:
ValueError: array item of unknown size: 'struct pollfd'

CFFI has no idea what size struct pollfd is, teach it the size. Note
this will fail horribly on non-linux systems, but currently we
conditional handle the code only for linux systems. However, this will
need to be addressed in the future for other platforms.

Signed-off-by: William Roberts <william.c.roberts@intel.com>