types: make TSS2_RC it's own class
Move the constants from TSS2_Exception to it's own class that matches
the spec's TSS2_RC and create a base class that contains a decode
routine for both TPM2_RC and TSS2_RC's. Use the decode function in
TSS2_Exception for decoding the error message and coerce int's into the
all inclusive TSS2_Exception type.

We have to be careful to not introduce circular deps, so we don't import
TSS2_RC type until runtime. Note that PEP563 also has a defered type
annotation mechanism that can be used now via from __future__ import
annotation. However, this is a python 3.7+ feature, and we don'e want to
use that. So for now, the type hints use the string name of the class.

Signed-off-by: William Roberts <william.c.roberts@intel.com>