constants/types: update docs and exceptions
Update the exceptions that the various parse routines can throw to use
ValueError and TypeError over the less specific RuntimeError and
document the inherited interfaces.

This is was kept as one commit since detangling it was more of a
headache than what it is worth (we wouldn't try and revert part of it).

Signed-off-by: William Roberts <william.c.roberts@intel.com>