docs: move COC to md so Github sees it
GitHub can find a CODE_OF_CONDUCT file under docs ONLY if it has no
extension or md extension, ie .rst is not supported [1].

By moving it to the GH found name, it can be retrieved through the
GitHub API and is shown for developers.

The other benefit of this PR, is that docs can now be written in md
if the full rst support is not needed.

1. Even though it works for README.rst files.

Signed-off-by: William Roberts <william.c.roberts@intel.com>