ci: quote the minimal version in pip command line
In shell script, using `... pip>21.3` means "run the command and write
the output in a file named `21.3`. This leads to empty output in GitHub
Actions, such as in https://github.com/tpm2-software/tpm2-pytss/runs/4703569920?check_suite_focus=true#step:4:2004

Fix this by quoting the parameter, which makes the shell interpret it as
a string instead of an output redirection.

Signed-off-by: Nicolas Iooss <nicolas.iooss@ledger.fr>