ESAPI: fix double-close issue
When `ESAPI.close` is called twice, it could raise an exception:

    >>> from tpm2_pytss import *
    >>> ectx = ESAPI("tabrmd:bus_type=system")
    >>> ectx.close()
    >>> ectx.close()
    WARNING:esys:src/tss2-esys/esys_context.c:114:Esys_Finalize() Finalizing NULL context.
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
      File "/workspace/tpm2-pytss/tpm2_pytss/ESAPI.py", line 155, in close
        self._tcti.close()

Fix this by ensuring that `self._tcti` is not `None` before calling
`close`. While at it, do not call `Esys_Finalize` twice.

Fix a bug in the handling of the logic to close the TCTI in
ESAPI.close(). If a TCTI object is passed to ESAPI(), the close() method
on the TCTI is never invoked and the connection is never terminated,
thus subsequent calls will block as the TPM resource is still held.

Update the get_tcti or property tcti to not invoke the Esys_GetTcti
function.

Signed-off-by: Nicolas Iooss <nicolas.iooss@ledger.fr>
Signed-off-by: William Roberts <william.c.roberts@intel.com>