Make the build reproducible
Whilst working on the Reproducible Builds effort [0] I noticed that
tpm2-pytss could not be built reproducibly.

This is due to the documentation embedding the current build date. This PR
results in the build system using a date seeded from the `SOURCE_DATE_EPOCH`
environment variable... if it exists [1].

I originally filed this in Debian as bug #1022777 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/1022777

Signed-off-by: Chris Lamb <lamby@debian.org>