test: ensure that simulator processes are terminated/killed
Seems that we leak simulator processes if the simulator is not ready
within about one second, so kill the subprocess if it's not ready.

Also ensure the simulator is killed in test teardown to avoid building
up a set of processes which might slow down the build.

This should solve the issue where the parts of the CI flow which run
after the tests runs very slow which is due to swtpm processes running
in the background using the CPU and only leaving ~10% of the CPU to the
CI processes.

Signed-off-by: Erik Larsson <who+github@cnackers.org>