ci: minimize tss versions
The ci matrix is exploding in size, thus prune the tested version to
minimize the matrix.

Drop:
  - 2.4.6
  - 3.0.0
  - 3.0.3
Add:
  - 3.2.0

We add 3.2.0 as it's most recent and repology.org has it as the most
popular packaged version at 71 hits and contains a lot of the backport
fixes on older versions. We check version numbers in the Python code.

We keep 2.4.0 as it's the minimum supported version.
We keep 3.1.0 as it checks that major release and misses a lot of the
backport fixes.

Fixes: #381

Signed-off-by: William Roberts <william.c.roberts@intel.com>