fapi: cleanup callbacks
The callbacks were implmented in a fashion where their was 10 versions
of each callback function in C and each version would get used in a
set_<fn>_callback routine. Ie the first to set would get assigned
the C callback ending in 0, the next caller would get index 1, and
so forth until the index's were exhausted. This placed in artificial
limitation on the number of callback users. Instead, have a single
C function callback for ALL FAPI context callers and demux based on the
callable and userdata.

SetCallback functions associate an _FAPI_CB_UDATA object which contains:
1. The Python Callback Callable
2. The Userdata
3. Any exceptions from the Callable that need to be re-raised.

This is all stored in the userdata for the C callback. If None is passed
for the callback, the state is cleared and the C callback is set to NULL.

A Fapi Function may then invoke a callback, if set, the userdata will be
the _FAPI_CB_UDATA, which we can then use to get to the user defined
callback and userdata. We can also use this object to get exceptions in
the python callback back accross the C interface and re-raise them if
needed.

TODO:
- Test policy action callback unset
- Test set_sign_callback unset
- Test set_branch_callback unset

Signed-off-by: William Roberts <william.c.roberts@intel.com>