constants: add routine for making ESYS_TR from parts
Relating to the following PRs:
- tpm2-tools: https://github.com/tpm2-software/tpm2-tools/pull/3214
- systemd: https://github.com/williamcroberts/systemd/tree/add-esys-tr-code

Note: systemd is part of a long term strategy with Dan Street and
Lennart Pottering over internal representation for disk key parents.

Create a function that takes a persistent handle and TPM2B_PUBLIC and
outputs a serialized ESYS_TR. This allows the creation of stable ESYS_TR
formats when a TPM is not available.

Signed-off-by: William Roberts <william.c.roberts@intel.com>