FAPI: Restore environment when exiting FAPIConfig
The FAPIConfig context manager was not restoring environment variables to
the state prior to entering the context. A consequence of that was that it
was not possible to specify a config file using the TSS2_FAPICONF
environment variable while running tests. The first test would use it, but
then the variable would be cleared for subsequent tests, and it would
revert back to trying the default paths.

Signed-off-by: Scott Stephens <stephens.js@gmail.com>